requirements.txt
flask
flask_session
python-dotenv
spotipy
Pillow
colorthief
